+++ common/src/main/java/dev/wuffs/squatgrow/SquatGrow.java
@@ -1,9 +1,8 @@
 package dev.wuffs.squatgrow;
 
 import dev.architectury.event.events.common.LifecycleEvent;
-import dev.architectury.event.events.common.TickEvent;
-import dev.architectury.platform.Platform;
 import dev.architectury.registry.ReloadListenerRegistry;
+import dev.wuffs.squatgrow.actions.Actions;
 import dev.wuffs.squatgrow.config.SquatGrowConfig;
 import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.ConfigHolder;
@@ -17,8 +16,8 @@ import net.minecraft.tags.TagKey;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -26,10 +25,7 @@ import java.util.stream.Collectors;
 
 public class SquatGrow {
     public static final String MOD_ID = "squatgrow";
-    private static final Logger LOGGER = LogManager.getLogger();
-    public static Logger getLogger() {
-        return LOGGER;
-    }
+    private static final Logger LOGGER = LoggerFactory.getLogger(SquatGrow.class);
 
     public static SquatGrowConfig config;
     public static ConfigHolder<SquatGrowConfig> configHolder;
@@ -58,8 +54,8 @@ public class SquatGrow {
     }
 
     private static void onSetup() {
-        SquatAction.isMysticalLoaded = Platform.isModLoaded("mysticalagriculture");
-        SquatAction.isAE2Loaded = Platform.isModLoaded("ae2");
+        LOGGER.debug("Starting setup");
+        Actions.get().setup();
     }
 
     /**
@@ -87,56 +83,6 @@ public class SquatGrow {
 
     public static Boolean allowTwerk(BlockState state) {
         return config.useWhitelist == isBlockInIgnoreList(state);
-//        ResourceLocation resourceLocation = state.getBlock().arch$registryName();
-//        if (wildcardCache.contains(resourceLocation.getNamespace())) {
-//
-//        }
-//        var regName = Registry.BLOCK.getKey(state.getBlock()).toString();
-//        var rawTags = state.getBlock().builtInRegistryHolder().tags().toList();
-//        if (config.useWhitelist) {
-//            if (config.ignorelist.contains(regName)) {
-//                if (config.debug) {
-//                    LOGGER.debug("Whitelist contains reg: " + regName);
-//                }
-//                return true;
-//            }
-//            if (rawTags.size() > 0) {
-//                for (TagKey<Block> tag : rawTags) {
-//                    if (config.ignorelist.contains("#" + tag.location().toString())) {
-//                        if (config.debug) {
-//                            LOGGER.debug("Whitelist contains tag: " + "#" + tag.location());
-//                        }
-//                        return true;
-//                    }
-//                }
-//            }
-//            if (config.debug) {
-//                LOGGER.debug("Whitelist does NOT contain: " + regName);
-//            }
-//            return false;
-//        }
-//
-//        if (config.ignorelist.contains(regName)) {
-//            if (config.debug) {
-//                LOGGER.debug("Blacklist contains reg: " + regName);
-//            }
-//            return false;
-//        }
-//
-//        if (rawTags.size() > 0) {
-//            for (TagKey<Block> tag : rawTags) {
-//                if (config.ignorelist.contains("#" + tag.location().toString())) {
-//                    return false;
-//                }
-//                if (config.debug) {
-//                    LOGGER.debug("Blacklist does NOT contain tag: " + "#" + tag.location());
-//                }
-//            }
-//        }
-//        if (config.debug) {
-//            LOGGER.debug("Nothing contains: " + regName);
-//        }
-//        return true;
     }
 
     private static boolean isBlockInIgnoreList(BlockState state) {